Removed create assistant page: we create the assistant, and then edit it
