+++ logicle/app/api/files/route.ts
@@ -1,14 +1,13 @@
 import { nanoid } from 'nanoid'
 import ApiResponses from '../utils/ApiResponses'
 import { requireSession } from '../utils/auth'
-import { InsertableFile } from '@/types/dto'
+import * as dto from '@/types/dto'
 import { addFile } from 'models/file'
-import { db } from '@/db/database'
 import Assistants from 'models/assistant'
 
 export const POST = requireSession(async (session, req) => {
   const id = nanoid()
-  const file = (await req.json()) as InsertableFile
+  const file = (await req.json()) as dto.InsertableFile
   const assistantId = req.nextUrl.searchParams.get('assistantId')
   const path = `${id}-${file.name.replace(/(\W+)/gi, '-')}`
   const created = await addFile(id, file, path)
